+++ R/f_bilateral_indicators.R
@@ -13,7 +13,7 @@
 #' @references
 #' {Bennet, T. L. (1920). \emph{The Theory of Measurement of Changes in Cost of Living}. Journal of the Royal Statistical Society, 83, 455-462.}
 #'
-#' {Białek, J. (2024). \emph{The use of the Bennet indicators and their transitive versions for scanner data analysis}. Statistics in Transition new series, 25(3).}
+#' {Białek, J. (2024). \emph{The use of the Bennet indicators and their transitive versions for scanner data analysis}. Statistics in Transition new series, 25(3), 155-173.}
 #' @examples 
 #' bennet(milk, "2018-12", "2019-12", matched=TRUE, contributions=TRUE)
 #' \donttest{bennet(coffee, start="2018-12", end="2019-03", interval=TRUE)}